+++ sdk/core/System.ClientModel/CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Other Changes
 
 - Removed `ReturnWhen` enum in favor of using bool `waitUntilCompleted` parameter in third-party client LRO method signatures.
+- Added abstract `UpdateStatus` method to `OperationResult`.
 
 ## 1.1.0-beta.6 (2024-08-01)