+++ sdk/core/System.ClientModel/CHANGELOG.md
@@ -4,11 +4,16 @@
 
 ### Features Added
 
+- Added `JsonModelConverter` to allow integration with System.Text.Json.
+
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+- Removed `ReturnWhen` enum in favor of using bool `waitUntilCompleted` parameter in third-party client LRO method signatures.
+- Added abstract `UpdateStatus` method to `OperationResult`.
 
 ## 1.1.0-beta.6 (2024-08-01)