+++ sdk/core/System.ClientModel/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 1.1.0-beta.8 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 1.1.0-beta.7 (2024-08-14)
 
 ### Features Added